chore: reorganize routes
