+++ app/routes/admin/logout.tsx
