+++ app/routes/projects/[projectId].tsx
@@ -3,10 +3,10 @@ import React from "react";
 import { Link, data, useOutletContext } from "react-router";
 import { ConditionalRichTextRenderer } from "~/routes/common/components/ConditionalRichTextRenderer";
 import { FadeIn } from "~/routes/common/components/ui/FadeIn";
-import type { loader as parentLayoutLoader } from "~/routes/projects/views/_layout";
+import type { loader as parentLayoutLoader } from "~/routes/projects/layout";
 import { assert } from "~/utils/assert";
-import { fetchProjectDetails } from "../services";
 import type { Route } from "./+types/[projectId]";
+import { fetchProjectDetails } from "./services";
 
 export const loader = async ({
 	params,
